requirements.txt
streamlit
pandas
plotly
scikit-learn
google-genai
python-dotenv
openpyxl
numpy